+++ src/index.ts
@@ -12,12 +12,15 @@ export {
   LoginError,
   NoActiveSessionError,
   StartVerificationTimeoutError,
+  UnresolvedSessionError,
+  VehicleNotReadyError,
 } from './exceptions.js';
 
 export type {
   Account,
   AccountBalance,
   APIEndpoints,
+  ChargePointCommandErrorBody,
   ChargePointOptions,
   ChargeSchedule,
   ChargingStatus,